packages/ai-gateway/pyproject.toml
[tool.poetry]
name = "ai-gateway"
version = "0.1.0"
description = ""
authors = ["VpgMackan <dev@mndev.se>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.95.0"
pydantic-settings = "^2.10.1"
httpx = "^0.28.1"
fastapi = "^0.116.1"
uvicorn = { extras = ["standard"], version = "^0.35.0" }
jinja2 = "^3.1.6"
structlog = "^24.1.0"
asgi-correlation-id = "^4.0.1"
elevenlabs = "^1.0.0"
groq = "^0.31.1"
posthog = "^7.4.3"
opentelemetry-distro = "^0.60b1"
opentelemetry-exporter-otlp = "^1.39.1"
opentelemetry-instrumentation-fastapi = "^0.60b1"
opentelemetry-instrumentation-httpx = "^0.60b1"


[tool.poetry.scripts]
gateway = "ai_gateway.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
